feat(localize): move parseNumber to @lion/localize
