+++ .changeset/dry-clocks-notice.md
@@ -1,0 +1,5 @@
+---
+'@lion/localize': minor
+---
+
+added a parseNumber function
