+++ packages/input-amount/index.js
@@ -1,4 +1,4 @@
 export { LionInputAmount } from './src/LionInputAmount.js';
 export { formatAmount } from './src/formatters.js';
-export { parseAmount, parseNumber } from './src/parsers.js';
+export { parseAmount } from './src/parsers.js';
 export { preprocessAmount } from './src/preprocessors.js';
