+++ packages/localize/index.js
@@ -4,13 +4,14 @@ export { getMonthNames } from './src/date/getMonthNames.js';
 export { getWeekdayNames } from './src/date/getWeekdayNames.js';
 export { normalizeDateTime } from './src/date/normalizeDateTime.js';
 export { parseDate } from './src/date/parseDate.js';
+export { localize, setLocalize } from './src/localize.js';
+export { LocalizeManager } from './src/LocalizeManager.js';
+export { LocalizeMixin } from './src/LocalizeMixin.js';
 export { formatNumber } from './src/number/formatNumber.js';
 export { formatNumberToParts } from './src/number/formatNumberToParts.js';
 export { getCurrencyName } from './src/number/getCurrencyName.js';
 export { getDecimalSeparator } from './src/number/getDecimalSeparator.js';
 export { getFractionDigits } from './src/number/getFractionDigits.js';
 export { getGroupSeparator } from './src/number/getGroupSeparator.js';
-export { localize, setLocalize } from './src/localize.js';
-export { LocalizeManager } from './src/LocalizeManager.js';
-export { LocalizeMixin } from './src/LocalizeMixin.js';
 export { normalizeCurrencyLabel } from './src/number/normalizeCurrencyLabel.js';
+export { parseNumber } from './src/number/parseNumber.js';
